[PIR #882] Extract gitignore helpers out of scaffold.ts
